pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.64.0",
    "mcp>=1.13.1",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
]
